requirements.txt
# requirements.txt
dspy-ai>=2.4.0
openai>=1.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0
matplotlib>=3.6.0
pandas>=2.0.0
requests>=2.28.0